Add grunt-contrib-copy task
Need to copy over static assets that are not compiled/concatenated with
other JS/CSS assets. Why?

Need to be able to use IE Conditional Comments to add HTML5 shim. In
order to conditionally load the asset and not burden the other
upstanding citizens of the web, we need to keep HTML5 shim from being
compiled with the app js.

Enter grunt-contrib-copy, which just copies the 'static/' directory
directly into generated and dist.