Better header margin spacing
Banner-to-Header spacing was not consistent between video/no-video for
each of the responsive breakpoints.

Set the font-size in px at the banner header level
  H1s set the font-size at 1em (anchored off the parent header)
Set the header margin in ems so that the spacing is relative to the
font-size, regardless of screen width. (Separate em values are used
between video and non-video posts)

Lastly, default the posts to non-video (no classname needed). And since
the .video class is already taken, switch to a subjectively semantic
classname 'screencast'.